Phase 7: relocate timeline render/interaction modules
Move axis.ts, card.ts (→ cards.ts), session-band.ts (→ session-bands.ts),
and the colocated test into timeline/render/. Move zoom.ts (math + types)
into timeline/interactions/.

This is a mechanical relocation to match the layout the slice's AGENTS.md
and add-timeline-feature skill have always described — no logic changes.
The follow-on commits extract pan, reschedule, and quick-add-zones from
app.ts into timeline/interactions/.
